refactor: dedupe trackManagement parallel copy (phase 1.4)
