+++ packages/bot/src/handlers/player/trackHandlers.spec.ts
@@ -45,7 +45,7 @@ jest.mock('@lucky/shared/services', () => ({
     },
 }))
 
-jest.mock('../../utils/music/trackManagement/queueOperations', () => ({
+jest.mock('../../utils/music/queueOperations', () => ({
     replenishQueue: (...args: unknown[]) => replenishQueueMock(...args),
 }))
