+++ packages/bot/src/handlers/player/trackHandlers.ts
@@ -2,7 +2,7 @@ import type { Track, GuildQueue } from 'discord-player'
 import { QueueRepeatMode } from 'discord-player'
 import { infoLog, debugLog, errorLog, warnLog } from '@lucky/shared/utils'
 import { addTrackToHistory } from '../../utils/music/duplicateDetection'
-import { replenishQueue } from '../../utils/music/trackManagement/queueOperations'
+import { replenishQueue } from '../../utils/music/queueOperations'
 import { resetAutoplayCount } from '../../utils/music/autoplayManager'
 import { featureToggleService } from '@lucky/shared/services'
 import { constants } from '@lucky/shared/config'
