--- packages/bot/src/utils/music/trackManagement/queueOperations.ts
@@ -1,9 +1,0 @@
-export {
-    addTracksToQueue,
-    addTrackToQueue,
-    clearQueue,
-    shuffleQueue,
-    removeTrackFromQueue,
-    moveTrackInQueue,
-    replenishQueue,
-} from '../queueOperations'
