--- packages/bot/src/utils/music/trackManagement/service.ts
@@ -1,1 +1,0 @@
-export { TrackManagementService } from '../service'
