--- packages/bot/src/utils/music/trackManagement/trackValidator.ts
@@ -1,1 +1,0 @@
-export { validateTrack, validateTracks } from '../trackValidator'
